switched to using unsigned 64 bit integers

+++ dynamic_programming/fib.cpp
@@ -1,18 +1,18 @@
 #pragma once 
 #include <iostream>
 
-int recursiveFib(int n) {
+uint64_t recursiveFib(uint64_t n) {
    if (n <= 1) return n;
    return recursiveFib(n - 1) + recursiveFib(n - 2);
 }
 
-int iterativeFib(int n) {
+uint64_t iterativeFib(uint64_t n) {
    if (n <= 1) return n;
 
-   int a = 0;
-   int b = 1;
-   int c = a + b;
-   for (int i = 2; i < n; i++) {
+   uint64_t a = 0;
+   uint64_t b = 1;
+   uint64_t c = a + b;
+   for (uint64_t i = 2; i < n; i++) {
       a = b;
       b = c;
       c = a + b;
